ut for devset; draft for idle queue

+++ ipshark2/core/devset.h
@@ -2,7 +2,7 @@
 #define  __DEVSET_H__
 #include "intern.h"
 
-void *__devset_create(bool black, size_t elenum, size_t size);
+void *__devset_create(bool black, int elenum, int size);
 void __devset_order(void);
 bool devset_ignore(char *s);
 void devset_exit(void);
@@ -10,7 +10,7 @@ void devset_exit(void);
 #define devset_init(l)							\
 ({									\
 	int err = 0;							\
-	size_t len = 0;							\
+	int len = 0;							\
 	bool black = true;						\
 	void *kp = NULL;						\
 	do {								\
